homework 5 - FULL
